+++ interviewPreparation/src/objectOrientedDesign/bankApplication/Util/BankConstants.java
@@ -9,8 +9,8 @@ package objectOrientedDesign.bankApplication.Util;
  * @version $Id: BankConstants.java, v 0.1 2020-05-19 13:15 rahul.jaiman Exp $$
  */
 public class BankConstants {
-    public static final double CURRENT_ACCOUNT_MINIMUM_BALANCE = 20000.0;
-    public static final double SAVING_ACCOUNT_MINIMUM_BALANCE = 10000.0;
+    public static final long CURRENT_ACCOUNT_MINIMUM_BALANCE = 20000;
+    public static final long SAVING_ACCOUNT_MINIMUM_BALANCE = 10000;
     public static final double CURRENT_ACCOUNT_INTEREST_RATE = 0.0;
     public static final double SAVING_ACCOUNT_INTEREST_RATE = 4.5;
     public static final String ZETA_BANK_ACCOUNT_NUMBER_PREFIX = "ZBAN";
